profil hotela crud

+++ booking/src/main/java/hotel/HotelRepository.java
@@ -1,8 +1,10 @@
 package hotel;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+
+
 import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface HotelRepository extends JpaRepository<Hotel, Integer> {
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
 }
