+++ booking/src/main/java/soba/SobaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SobaRepository extends JpaRepository<Soba, Integer>{
+public interface SobaRepository extends JpaRepository<Soba, Long>{
 }
